refactor packages and remove useless class
