+++ bandyer-android-design/src/main/java/com/bandyer/sdk_design/filesharing/FileShareViewModel.kt
@@ -3,7 +3,7 @@ package com.bandyer.sdk_design.filesharing
 import android.content.Context
 import android.net.Uri
 import androidx.lifecycle.ViewModel
-import java.io.File
+import com.bandyer.sdk_design.filesharing.model.FileShareItemData
 import java.util.concurrent.ConcurrentHashMap
 
 abstract class FileShareViewModel : ViewModel() {
