+++ bandyer-android-design/src/main/java/com/bandyer/sdk_design/filesharing/adapter_items/DownloadAvailableItem.kt
