+++ bandyer-android-design/src/main/java/com/bandyer/sdk_design/filesharing/imageviews/BandyerFileShareOpTypeImageView.kt
@@ -1,4 +1,4 @@
-package com.bandyer.sdk_design.filesharing
+package com.bandyer.sdk_design.filesharing.imageviews
 
 import android.content.Context
 import android.util.AttributeSet
