+++ bandyer-android-design/src/main/java/com/bandyer/sdk_design/filesharing/model/DownloadItemData.kt
@@ -1,4 +1,4 @@
-package com.bandyer.sdk_design.filesharing
+package com.bandyer.sdk_design.filesharing.model
 
 interface DownloadItemData: FileShareItemData {
     val endpoint: String
