+++ bandyer-android-design/src/main/java/com/bandyer/sdk_design/filesharing/model/UploadItemData.kt
@@ -1,4 +1,4 @@
-package com.bandyer.sdk_design.filesharing
+package com.bandyer.sdk_design.filesharing.model
 
 import android.net.Uri
